[FIX] Clean code & run all features of users & groups
